:construction: semaforos :construction:

+++ RowDaBoat-x64barebones-d4e1c147f975/Userland/SampleCodeModule/include/semLib.h
@@ -4,10 +4,10 @@
 #include <stdint.h>
 
 int sem_open(char * name, uint64_t value);
-int sem_wait(uint64_t index);
-int sem_post(uint64_t index);
-int sem_close(uint64_t index);
-void sem_changeValue(uint64_t index, uint64_t value);
+int sem_wait(int index);
+int sem_post(int index);
+int sem_close(int index);
+void sem_changeValue(int index, uint64_t value);
 void sem_print();
 
 #endif
